requirements.txt
# Video Archive Tool - Python Dependencies
# Python 3.11+ required

# Core Video Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
scenedetect[opencv]==0.6.2
ffmpeg-python==0.2.0

# Image Processing
Pillow==10.1.0
numpy==1.24.3

# GUI Framework
# tkinter is included with Python

# Database
# sqlite3 is included with Python

# Metadata Handling
piexif==1.1.3
mutagen==1.47.0

# Utilities
psutil==5.9.6  # System monitoring
colorlog==6.7.0  # Colored logging
python-dotenv==1.0.0  # Environment variables
tqdm==4.66.1  # Progress bars

# Development Dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0

# Build Dependencies
pyinstaller==6.2.0
auto-py-to-exe==2.41.0  # Optional GUI for PyInstaller